requirements.txt
Flask==2.3.3
pyodbc==4.0.39
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
pytest==7.4.2
pytest-flask==1.2.0
Werkzeug==2.3.7
Jinja2==3.1.2